Git commit message:
feat: Add mappings for new and Spectre in user/mappings.lua

In this commit, we added mapping functions for `n` and `<leader>s` in
user/mappings.lua. We also removed the `is_available` check from the
`nvim-dap` section of the code. Finally, we added a comment section
to give more context to the code in user/polish.lua.

Co-authored-by: [insert co-author name and email] (if applicable)